Check draw tool exists before delete

+++ x-pack/plugins/maps/public/connected_components/mb_map/draw_control/draw_feature_control/draw_feature_control.tsx
@@ -88,7 +88,9 @@ export class DrawFeatureControl extends Component<Props, {}> {
       );
     } finally {
       this.props.disableDrawState();
-      mbDrawControl.deleteAll();
+      if (mbDrawControl) {
+        mbDrawControl.deleteAll();
+      }
     }
   };
